Utils: Fix circular dependency caused by separation
When separating Kano Utils into submodules, a circular dependency
between the `shell.py` and `process.py` modules was created. Resolve
this by migrating the offending code to be contained in just the `shell`
module.
