+++ kano/utils/shell.py
@@ -1,8 +1,14 @@
 import os
 import sys
+import signal
 import subprocess
 
-from kano.utils.processes import restore_signals
+def restore_signals():
+    signals = ('SIGPIPE', 'SIGXFZ', 'SIGXFSZ')
+    for sig in signals:
+        if hasattr(signal, sig):
+            signal.signal(getattr(signal, sig), signal.SIG_DFL)
+
 
 def run_cmd(cmd, localised=False, unsudo=False):
     '''
